Update mysql_user.py
There are two lines that cause a `TypeError: not all arguments converted during string formatting` because the single argument is not a tuple, in `python 2.7.11` with `1.3.7-1build2`.

This commit also reverts 23dd16c6 (PR #3038), as I feel it tries to solve the same problem, being that cursor.execute expects a tuple.